Move tests into __tests__
Some tests were in src/. This moves them into src/__tests__/ and adjusts
the imports and mocks accordingly.
