+++ src/components/__tests__/Home.test.tsx
@@ -1,12 +1,13 @@
-import React from 'react';
 import { shallow } from 'enzyme';
-import Home from './Home';
+import React from 'react';
+import Home from '../Home';
+
 jest.mock('react-router-dom/Link');
 jest.mock('@material-ui/core/Button', () => 'button-material-ui');
 jest.mock('@material-ui/core/Grid', () => 'Grid');
-jest.mock('./Icons');
-jest.mock('./SearchInput', () => 'SearchInput');
-jest.mock('./HomeButtons', () => 'HomeButtons');
+jest.mock('../Icons');
+jest.mock('../SearchInput', () => 'SearchInput');
+jest.mock('../HomeButtons', () => 'HomeButtons');
 
 describe('<Home/>', () => {
   const wrapper = shallow(<Home />);
